better encapsulation: moving types to the modules in which they're used
