+++ src/main.rs
@@ -1,193 +1,15 @@
-use std::fmt::Formatter;
-
 use bevy::asset::AssetMetaCheck;
 use bevy::ecs::system::EntityCommands;
 use bevy::prelude::*;
-use dimension_macro_derive::Dimension;
-use rand::prelude::*;
 
 mod menu;
 mod game;
 
-#[derive(Debug, PartialEq, Eq, Hash, Clone, Copy, Dimension, Component)]
-enum Row {
-    Bottom,
-    Middle,
-    Top,
-}
-
-#[derive(Debug, PartialEq, Eq, Hash, Clone, Copy, Dimension, Component)]
-enum Column {
-    Left,
-    Middle,
-    Right
-}
-
-#[derive(Clone, Copy)]
-enum Line {
-    BottomRow,
-    MiddleRow,
-    TopRow,
-    LeftColumn,
-    MiddleColumn,
-    RightColumn,
-    UpDiagonal,
-    DownDiagonal,
-}
-
-impl Into<[Cell;3]> for Line {
-    fn into(self) -> [Cell; 3] {
-        match self {
-            Self::BottomRow => [Cell::BottomLeft, Cell::BottomMiddle, Cell::BottomRight],
-            Self::MiddleRow => [Cell::MiddleLeft, Cell::MiddleMiddle, Cell::MiddleRight],
-            Self::TopRow => [Cell::TopLeft, Cell::TopMiddle, Cell::TopRight],
-            Self::LeftColumn => [Cell::TopLeft, Cell::MiddleLeft, Cell::BottomLeft],
-            Self::MiddleColumn => [Cell::TopMiddle, Cell::MiddleMiddle, Cell::BottomMiddle],
-            Self::RightColumn => [Cell::TopRight, Cell::MiddleRight, Cell::BottomRight],
-            Self::UpDiagonal => [Cell::BottomLeft, Cell::MiddleMiddle, Cell::TopRight],
-            Self::DownDiagonal => [Cell::TopLeft, Cell::MiddleMiddle, Cell::BottomRight],
-        }
-    }
-}
-
-impl Line {
-    fn all() -> [Self;8] {
-        [
-            Self::BottomRow,
-            Self::MiddleRow,
-            Self::TopRow,
-            Self::LeftColumn,
-            Self::MiddleColumn,
-            Self::RightColumn,
-            Self::UpDiagonal,
-            Self::DownDiagonal,
-        ]
-    }
-}
-
-#[derive(Component, PartialEq, Eq, Hash, Clone, Copy, Debug)]
-enum Cell {
-    TopLeft,
-    TopMiddle,
-    TopRight,
-    MiddleLeft,
-    MiddleMiddle,
-    MiddleRight,
-    BottomLeft,
-    BottomMiddle,
-    BottomRight,
-}
-
-impl Cell {
-    fn all() -> [Self;9] {
-        [
-            Self::TopLeft,
-            Self::TopMiddle,
-            Self::TopRight,
-            Self::MiddleLeft,
-            Self::MiddleMiddle,
-            Self::MiddleRight,
-            Self::BottomLeft,
-            Self::BottomMiddle,
-            Self::BottomRight,
-        ]
-    }
-
-    fn row(&self) -> Row {
-        match self {
-            Cell::TopLeft => Row::Top,
-            Cell::TopMiddle => Row::Top,
-            Cell::TopRight => Row::Top,
-            Cell::MiddleLeft => Row::Middle,
-            Cell::MiddleMiddle => Row::Middle,
-            Cell::MiddleRight => Row::Middle,
-            Cell::BottomLeft => Row::Bottom,
-            Cell::BottomMiddle => Row::Bottom,
-            Cell::BottomRight => Row::Bottom,
-        }
-    }
-
-    fn column(&self) -> Column {
-        match self {
-            Cell::TopLeft => Column::Left,
-            Cell::TopMiddle => Column::Middle,
-            Cell::TopRight => Column::Right,
-            Cell::MiddleLeft => Column::Left,
-            Cell::MiddleMiddle => Column::Middle,
-            Cell::MiddleRight => Column::Right,
-            Cell::BottomLeft => Column::Left,
-            Cell::BottomMiddle => Column::Middle,
-            Cell::BottomRight => Column::Right,
-        }
-    }
-
-    fn from(row: Row, column: Column) -> Cell {
-        match row {
-            Row::Bottom => match column {
-                Column::Left => Cell::BottomLeft,
-                Column::Middle => Cell::BottomMiddle,
-                Column::Right => Cell::BottomRight,
-            }
-            Row::Middle => match column {
-                Column::Left => Cell::MiddleLeft,
-                Column::Middle => Cell::MiddleMiddle,
-                Column::Right => Cell::MiddleRight,
-            }
-            Row::Top => match column {
-                Column::Left => Cell::TopLeft,
-                Column::Middle => Cell::TopMiddle,
-                Column::Right => Cell::TopRight,
-            }
-        }
-    }
-
-    fn is_corner(&self) -> bool {
-        *self == Self::TopLeft || *self == Self::TopRight || *self == Self::BottomLeft || *self == Self::BottomRight
-    }
-}
-
-struct Grid {}
-
-impl Grid {
-    fn hit_square(pos: Vec2) -> Option<Cell> {
-        match (Row::containing(pos.y), Column::containing(pos.x)) {
-            (None, _) | (_, None) => None,
-            (Some(row), Some(col)) => Some(Cell::from(row, col))
-        }
-    }
-}
-
-#[derive(Clone, Copy, Default, Eq, PartialEq, Debug, Hash, States, Component)]
+#[derive(States, Component, Clone, Copy, Debug, Default, PartialEq, Eq, Hash)]
 enum AppState {
     #[default]
     Menu,
     Game,
-}
-
-// when used as a Resource, Mark is the human player (the other player is the computer)
-#[derive(Component, Default, PartialEq, Eq, Debug, Clone, Copy, Hash)]
-enum Mark {
-    #[default]
-    X,
-    O
-}
-
-impl std::fmt::Display for Mark {
-    fn fmt(&self, f: &mut Formatter<'_>) -> std::fmt::Result {
-        match self {
-            Mark::X => write!(f, "X"),
-            Mark::O => write!(f, "O"),
-        }
-    }
-}
-
-impl Mark {
-    fn color(&self) -> Color {
-        match self {
-            Mark::X => Color::RED,
-            Mark::O => Color::BLUE,
-        }
-    }
 }
 
 #[derive(Resource, Component, Clone, Copy, Debug, Default, PartialEq, Eq)]
@@ -195,16 +17,6 @@ enum HumanMark {
     #[default]
     HumanX,
     HumanO
-}
-
-impl HumanMark {
-    fn is(&self, mark: Mark) -> bool {
-        match self {
-            HumanMark::HumanX if mark == Mark::X => true,
-            HumanMark::HumanO if mark == Mark::O => true,
-            _ => false
-        }
-    }
 }
 
 #[derive(Resource, Component, Clone, Copy, Debug, Default, PartialEq, Eq)]
